+++ app/presenters/checkboxes_answer_presenter.rb
@@ -1,0 +1,5 @@
+class CheckboxesAnswerPresenter < SimpleDelegator
+  def response
+    super.reject(&:blank?).map(&:capitalize).join(", ")
+  end
+end