+++ config/environments/test.rb
@@ -63,5 +63,6 @@ Rails.application.configure do
     Bullet.add_whitelist type: :unused_eager_loading, class_name: "Step", association: :short_text_answer
     Bullet.add_whitelist type: :unused_eager_loading, class_name: "Step", association: :long_text_answer
     Bullet.add_whitelist type: :unused_eager_loading, class_name: "Step", association: :single_date_answer
+    Bullet.add_whitelist type: :unused_eager_loading, class_name: "Step", association: :checkbox_answers
   end
 end