+++ db/migrate/20210118110545_change_options_to_jsonb.rb
@@ -1,0 +1,6 @@
+class ChangeOptionsToJsonb < ActiveRecord::Migration[6.1]
+  def change
+    remove_column :steps, :options
+    add_column :steps, :options, :jsonb
+  end
+end