+++ db/migrate/20210118123111_add_further_information_to_radio_answer.rb
@@ -1,0 +1,5 @@
+class AddFurtherInformationToRadioAnswer < ActiveRecord::Migration[6.1]
+  def change
+    add_column :radio_answers, :further_information, :text
+  end
+end