+++ spec/factories/answer.rb
@@ -3,6 +3,7 @@ FactoryBot.define do
     association :step, factory: :step, contentful_type: "radios", contentful_model: "question"
 
     response { "Green" }
+    further_information { nil }
   end
 
   factory :short_text_answer do
@@ -18,13 +19,13 @@ FactoryBot.define do
   end
 
   factory :single_date_answer do
-    association :step, factory: :step, contentful_type: "single_date"
+    association :step, factory: :step, contentful_type: "single_date", contentful_model: "question"
 
     response { 1.year.from_now }
   end
 
   factory :checkbox_answers do
-    association :step, factory: :step, contentful_type: "checkboxes"
+    association :step, factory: :step, options: [{"value" => "Breakfast"}, {"value" => "Lunch"}], contentful_type: "checkboxes", contentful_model: "question"
 
     response { ["breakfast", "lunch", ""] }
   end