+++ spec/factories/step.rb
@@ -2,13 +2,13 @@ FactoryBot.define do
   factory :step do
     title { "What is your favourite colour?" }
     help_text { "Choose the primary colour closest to your choice" }
-    raw { "{\"sys\":{\"id\"=>\"123\"}}" }
+    raw { {"sys": {"id" => "123"}} }
     contentful_id { "123" }
 
     association :journey, factory: :journey
 
     trait :radio do
-      options { ["Red", "Green", "Blue"] }
+      options { [{"value" => "Red"}, {"value" => "Green"}, {"value" => "Blue"}] }
       contentful_model { "question" }
       contentful_type { "radios" }
       association :radio_answer
@@ -36,7 +36,7 @@ FactoryBot.define do
     end
 
     trait :checkbox_answers do
-      options { ["Brown", "Gold"] }
+      options { [{"value" => "Brown"}, {"value" => "Gold"}] }
       contentful_model { "question" }
       contentful_type { "checkboxes" }
       association :checkbox_answers
@@ -46,7 +46,6 @@ FactoryBot.define do
       options { nil }
       contentful_model { "staticContent" }
       contentful_type { "paragraphs" }
-      association :paragraph_content
     end
   end
 end