+++ spec/features/visitors/anyone_can_complete_a_journey_spec.rb
@@ -3,35 +3,36 @@ require "rails_helper"
 feature "Anyone can start a journey" do
   around do |example|
     ClimateControl.modify(
-      CONTENTFUL_PLANNING_START_ENTRY_ID: "1UjQurSOi5MWkcRuGxdXZS"
+      CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-starting-step"
     ) do
       example.run
     end
   end
 
   scenario "Start page includes a call to action" do
-    stub_get_contentful_entry
+    stub_get_contentful_entries(
+      entry_id: "contentful-starting-step",
+      fixture_filename: "closed-path-with-multiple-example.json"
+    )
 
     visit root_path
 
     click_on(I18n.t("generic.button.start"))
 
-    expect(page).to have_content("Which service do you need?")
-    expect(page).to have_content("Tell us which service you need.")
     expect(page).to have_content("Catering")
-    expect(page).to have_content("Cleaning")
-
-    choose("Catering")
-
-    click_on(I18n.t("generic.button.next"))
+    expect(page).to have_content("Which service do you need?")
+    expect(page).to have_content("Not started")
   end
 
   scenario "an answer must be provided" do
-    stub_get_contentful_entry
+    stub_get_contentful_entries(
+      entry_id: "contentful-starting-step",
+      fixture_filename: "closed-path-with-multiple-example.json"
+    )
+    journey = CreateJourney.new(category: "catering").call
+    step = journey.steps.find_by(contentful_id: "contentful-radio-question")
 
-    visit root_path
-
-    click_on(I18n.t("generic.button.start"))
+    visit journey_step_path(journey, step)
 
     # Omit a choice
 
@@ -40,113 +41,81 @@ feature "Anyone can start a journey" do
     expect(page).to have_content("can't be blank")
   end
 
-  context "when the starter step has a next step" do
-    around do |example|
-      ClimateControl.modify(
-        CONTENTFUL_PLANNING_START_ENTRY_ID: "47EI2X2T5EDTpJX9WjRR9p"
-      ) do
-        example.run
-      end
-    end
-
-    scenario "there are 2 steps to answer" do
-      visit root_path
-
-      stub_get_contentful_entry(
-        entry_id: "47EI2X2T5EDTpJX9WjRR9p",
-        fixture_filename: "has-next-question-example.json"
-      )
-      click_on(I18n.t("generic.button.start"))
-
-      choose("Catering")
-
-      stub_get_contentful_entry(
-        entry_id: "5lYcZs1ootDrOnk09LDLZg",
-        fixture_filename: "no-next-question-example.json"
-      )
-
-      click_on(I18n.t("generic.button.next"))
-
-      choose("Stationary")
-      click_on(I18n.t("generic.button.next"))
-
-      expect(page).to have_content("Catering")
-      expect(page).to have_content("Stationary")
-    end
-  end
-
   context "when the Contentful model is of type question" do
     context "when Contentful entry is of type short_text" do
       around do |example|
         ClimateControl.modify(
-          CONTENTFUL_PLANNING_START_ENTRY_ID: "hfjJgWRg4xiiiImwVRDtZ"
+          CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-starting-step"
         ) do
           example.run
         end
       end
 
       scenario "user can answer using free text" do
-        stub_get_contentful_entry(
-          entry_id: "hfjJgWRg4xiiiImwVRDtZ",
-          fixture_filename: "short-text-question-example.json"
+        stub_get_contentful_entries(
+          entry_id: "contentful-starting-step",
+          fixture_filename: "closed-path-with-multiple-example.json"
         )
+        journey = CreateJourney.new(category: "catering").call
+        step = journey.steps.find_by(contentful_id: "contentful-short-text-question")
 
-        visit root_path
-        click_on(I18n.t("generic.button.start"))
+        visit journey_step_path(journey, step)
 
         fill_in "answer[response]", with: "email@example.com"
         click_on(I18n.t("generic.button.next"))
 
-        expect(page).to have_content("email@example")
+        click_on(step.title)
+
+        expect(find_field("answer-response-field").value).to eql("email@example.com")
       end
     end
 
     context "when Contentful entry is of type long_text" do
       around do |example|
         ClimateControl.modify(
-          CONTENTFUL_PLANNING_START_ENTRY_ID: "2jWIO1MrVIya9NZrFWT4e"
+          CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-starting-step"
         ) do
           example.run
         end
       end
 
       scenario "user can answer using free text with multiple lines" do
-        stub_get_contentful_entry(
-          entry_id: "2jWIO1MrVIya9NZrFWT4e",
-          fixture_filename: "long-text-question-example.json"
+        stub_get_contentful_entries(
+          entry_id: "contentful-starting-step",
+          fixture_filename: "closed-path-with-multiple-example.json"
         )
+        journey = CreateJourney.new(category: "catering").call
+        step = journey.steps.find_by(contentful_id: "contentful-long-text-question")
 
-        visit root_path
-        click_on(I18n.t("generic.button.start"))
+        visit journey_step_path(journey, step)
 
-        fill_in "answer[response]", with: "We would like a supplier to provide catering from September 2020.\r\nThey must be able to supply us for 3 years minumum."
+        fill_in "answer[response]", with: "We would like a supplier to provide catering from September 2020.\nThey must be able to supply us for 3 years minumum."
         click_on(I18n.t("generic.button.next"))
 
-        within(".govuk-summary-list") do
-          paragraphs_elements = find_all("p")
-          expect(paragraphs_elements.first.text).to have_content("We would like a supplier to provide catering from September 2020.")
-          expect(paragraphs_elements.last.text).to have_content("They must be able to supply us for 3 years minumum.")
-        end
+        click_on(step.title)
+
+        expect(find_field("answer-response-field").value).to eql("We would like a supplier to provide catering from September 2020.\r\nThey must be able to supply us for 3 years minumum.")
       end
     end
 
     context "when Contentful entry is of type single_date" do
       around do |example|
         ClimateControl.modify(
-          CONTENTFUL_PLANNING_START_ENTRY_ID: "55G5kpCLLL3h5yBQLiVlYy"
+          CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-starting-step"
         ) do
           example.run
         end
       end
 
       scenario "user can answer using a date input" do
-        stub_get_contentful_entry(
-          entry_id: "55G5kpCLLL3h5yBQLiVlYy",
-          fixture_filename: "single-date-example.json"
+        stub_get_contentful_entries(
+          entry_id: "contentful-starting-step",
+          fixture_filename: "closed-path-with-multiple-example.json"
         )
+        journey = CreateJourney.new(category: "catering").call
+        step = journey.steps.find_by(contentful_id: "contentful-single-date-question")
 
-        visit root_path
-        click_on(I18n.t("generic.button.start"))
+        visit journey_step_path(journey, step)
 
         fill_in "answer[response(3i)]", with: "12"
         fill_in "answer[response(2i)]", with: "8"
@@ -154,34 +123,78 @@ feature "Anyone can start a journey" do
 
         click_on(I18n.t("generic.button.next"))
 
-        expect(page).to have_content("12 Aug 2020")
+        click_on(step.title)
+
+        expect(find_field("answer_response_3i").value).to eql("12")
+        expect(find_field("answer_response_2i").value).to eql("8")
+        expect(find_field("answer_response_1i").value).to eql("2020")
       end
     end
 
     context "when Contentful entry is of type checkboxes" do
       around do |example|
         ClimateControl.modify(
-          CONTENTFUL_PLANNING_START_ENTRY_ID: "1DqhwF2XkJJ0Um6NSweWlZ"
+          CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-starting-step"
         ) do
           example.run
         end
       end
 
       scenario "user can select multiple answers" do
-        stub_get_contentful_entry(
-          entry_id: "1DqhwF2XkJJ0Um6NSweWlZ",
-          fixture_filename: "checkbox-example.json"
+        stub_get_contentful_entries(
+          entry_id: "contentful-starting-step",
+          fixture_filename: "closed-path-with-multiple-example.json"
         )
+        journey = CreateJourney.new(category: "catering").call
+        step = journey.steps.find_by(contentful_id: "contentful-checkboxes-question")
 
-        visit root_path
-        click_on(I18n.t("generic.button.start"))
+        visit journey_step_path(journey, step)
 
         check "Breakfast"
         check "Lunch"
 
         click_on(I18n.t("generic.button.next"))
 
-        expect(page).to have_content("Breakfast, Lunch")
+        click_on(step.title)
+
+        expect(page).to have_checked_field("answer-response-breakfast-field")
+        expect(page).to have_checked_field("answer-response-lunch-field")
+      end
+    end
+
+    context "when Contentful entry is of type radios" do
+      context "when extra configuration is passed to collect further info" do
+        around do |example|
+          ClimateControl.modify(
+            CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-starting-step"
+          ) do
+            example.run
+          end
+        end
+
+        scenario "asks the user for further information" do
+          stub_get_contentful_entries(
+            entry_id: "contentful-starting-step",
+            fixture_filename: "closed-path-with-multiple-example.json"
+          )
+          journey = CreateJourney.new(category: "catering").call
+          step = journey.steps.find_by(contentful_id: "contentful-radio-question")
+
+          visit journey_step_path(journey, step)
+
+          click_on(I18n.t("generic.button.start"))
+
+          choose("Catering")
+          fill_in "answer[further_information]", with: "The school needs the kitchen cleaned once a day"
+
+          click_on(I18n.t("generic.button.next"))
+
+          click_on(step.title)
+
+          expect(page).to have_checked_field("Catering")
+          expect(find_field("answer-further-information-field").value)
+            .to eql("The school needs the kitchen cleaned once a day")
+        end
       end
     end
   end
@@ -190,20 +203,21 @@ feature "Anyone can start a journey" do
     context "when Contentful entry is of type paragraphs" do
       around do |example|
         ClimateControl.modify(
-          CONTENTFUL_PLANNING_START_ENTRY_ID: "5kZ9hIFDvNCEhjWs72SFwj"
+          CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-starting-step"
         ) do
           example.run
         end
       end
 
       scenario "user can read static content and proceed without answering" do
-        stub_get_contentful_entry(
-          entry_id: "5kZ9hIFDvNCEhjWs72SFwj",
-          fixture_filename: "static-content-example.json"
+        stub_get_contentful_entries(
+          entry_id: "contentful-starting-step",
+          fixture_filename: "closed-path-with-multiple-example.json"
         )
+        journey = CreateJourney.new(category: "catering").call
+        step = journey.steps.find_by(contentful_id: "contentful-starting-step")
 
-        visit root_path
-        click_on(I18n.t("generic.button.start"))
+        visit journey_step_path(journey, step)
 
         expect(page).to have_content("When you should start")
 
@@ -223,21 +237,19 @@ feature "Anyone can start a journey" do
   context "when Contentful entry model wasn't an expected type" do
     around do |example|
       ClimateControl.modify(
-        CONTENTFUL_PLANNING_START_ENTRY_ID: "6EKsv389ETYcQql3htK3Z2"
+        CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-unexpected-model"
       ) do
         example.run
       end
     end
 
     scenario "returns an error message" do
-      stub_get_contentful_entry(
-        entry_id: "6EKsv389ETYcQql3htK3Z2",
-        fixture_filename: "an-unexpected-model-example.json"
+      stub_get_contentful_entries(
+        entry_id: "contentful-unexpected-model",
+        fixture_filename: "path-with-unexpected-model.json"
       )
 
-      visit root_path
-
-      click_on(I18n.t("generic.button.start"))
+      visit new_journey_path
 
       expect(page).to have_content(I18n.t("errors.unexpected_contentful_model.page_title"))
       expect(page).to have_content(I18n.t("errors.unexpected_contentful_model.page_body"))
@@ -247,39 +259,58 @@ feature "Anyone can start a journey" do
   context "when the Contentful Entry wasn't an expected step type" do
     around do |example|
       ClimateControl.modify(
-        CONTENTFUL_PLANNING_START_ENTRY_ID: "8as7df68uhasdnuasdf"
+        CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-unexpected-step-type"
       ) do
         example.run
       end
     end
 
     scenario "returns an error message" do
-      stub_get_contentful_entry(
-        entry_id: "8as7df68uhasdnuasdf",
-        fixture_filename: "an-unexpected-question-type-example.json"
+      stub_get_contentful_entries(
+        entry_id: "contentful-unexpected-step-type",
+        fixture_filename: "path-with-unexpected-step-type.json"
       )
 
-      visit root_path
-
-      click_on(I18n.t("generic.button.start"))
+      visit new_journey_path
 
       expect(page).to have_content(I18n.t("errors.unexpected_contentful_step_type.page_title"))
       expect(page).to have_content(I18n.t("errors.unexpected_contentful_step_type.page_body"))
     end
   end
 
-  scenario "a Contentful entry_id does not exist" do
-    contentful_client = stub_contentful_client
+  context "when the starting entry id doesn't exist" do
+    around do |example|
+      ClimateControl.modify(
+        CONTENTFUL_PLANNING_START_ENTRY_ID: "contentful-fake-entry-id"
+      ) do
+        example.run
+      end
+    end
 
-    allow(contentful_client).to receive(:entry)
-      .with(anything)
-      .and_raise(GetContentfulEntry::EntryNotFound.new("The following Contentful error could not be found: sss "))
+    scenario "a Contentful entry_id does not exist" do
+      stub_get_contentful_entries(
+        entry_id: "contentful-fake-entry-id",
+        fixture_filename: "closed-path-with-multiple-example.json"
+      )
 
-    visit root_path
+      visit new_journey_path
 
-    click_on(I18n.t("generic.button.start"))
+      expect(page).to have_content(I18n.t("errors.contentful_entry_not_found.page_title"))
+      expect(page).to have_content(I18n.t("errors.contentful_entry_not_found.page_body"))
+    end
+  end
 
-    expect(page).to have_content(I18n.t("errors.contentful_entry_not_found.page_title"))
-    expect(page).to have_content(I18n.t("errors.contentful_entry_not_found.page_body"))
+  context "when the Liquid template was invalid" do
+    it "raises an error" do
+      fake_liquid_template = File.read("#{Rails.root}/spec/fixtures/specification_templates/invalid.liquid")
+      allow_any_instance_of(FindLiquidTemplate).to receive(:file).and_return(fake_liquid_template)
+
+      visit root_path
+
+      click_on(I18n.t("generic.button.start"))
+
+      expect(page).to have_content(I18n.t("errors.specification_template_invalid.page_title"))
+      expect(page).to have_content(I18n.t("errors.specification_template_invalid.page_body"))
+    end
   end
 end