+++ spec/presenters/checkboxes_answer_presenter_spec.rb
@@ -1,0 +1,11 @@
+require "rails_helper"
+
+RSpec.describe CheckboxesAnswerPresenter do
+  describe "#response" do
+    it "returns all none nil values, capitalised as a comma separated string" do
+      step = build(:checkbox_answers, response: ["yes", "no", ""])
+      presenter = described_class.new(step)
+      expect(presenter.response).to eq("Yes, No")
+    end
+  end
+end