+++ spec/requests/entry_preview_spec.rb
@@ -5,12 +5,20 @@ RSpec.describe "Entry previews", type: :request do
     entry_id = "123"
     fake_journey = create(:journey)
     expect(Journey).to receive(:create)
-      .with(category: anything, next_entry_id: entry_id)
+      .with(category: anything)
       .and_return(fake_journey)
+
+    fake_get_contentful_entry = instance_double(Contentful::Entry)
+    allow_any_instance_of(GetContentfulEntry).to receive(:call)
+      .and_return(fake_get_contentful_entry)
+
+    fake_step = create(:step, :radio)
+    allow_any_instance_of(CreateJourneyStep).to receive(:call)
+      .and_return(fake_step)
 
     get "/preview/entries/#{entry_id}"
 
     expect(response).to have_http_status(:found)
-    expect(response).to redirect_to("/journeys/#{fake_journey.id}/steps/new")
+    expect(response).to redirect_to("/journeys/#{fake_journey.id}/steps/#{fake_step.id}")
   end
 end